refactor: EmailService를 common/mail/MailService로 이동
email 발송은 공통 인프라 관심사이므로 별도 email 패키지 대신
common/mail 패키지로 이동
